requirements.txt
azure-functions==1.18.0
psycopg2-binary==2.9.9
pyodbc==5.1.0
